# back/openapi.yaml
openapi: 3.0.3
info:
  title: IS Course Work Back API
  description: >-
    REST API that powers the restaurant project. It exposes
    endpoints to manage employees, dishes, ingredients, tables, orders, bills,
    feedback, reports, and employee wallets. Most operations require a JWT that
    encodes the employee id and position (role).
  version: 1.0.0
servers:
  - url: http://localhost:8080/is_coursework/api
    description: Local application server
...
paths:
...
  /report/main/{dateStr}:
    get:
      tags: [Reports]
      summary: Summary report for a date
      description: Only administrators may view global reports.
      security:
        - BearerAuth: []
      parameters:
        - name: dateStr
          in: path
          description: ISO-8601 date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MainReportData'
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /report/employee-all/{dateStr}:
    get:
      tags: [Reports]
      summary: All employees report for a date
      description: Only administrators may view per-employee metrics.
      security:
        - BearerAuth: []
      parameters:
        - name: dateStr
          in: path
          description: ISO-8601 date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Employee report map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeReportMap'
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /report/employee-own/{dateStr}:
    get:
      tags: [Reports]
      summary: Employee self-report for a date
      description: Requires authentication; the report is calculated for the current token holder.
      security:
        - BearerAuth: []
      parameters:
        - name: dateStr
          in: path
          description: ISO-8601 date (YYYY-MM-DD)
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Employee report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeReportData'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /wallet/get-balance/{employeeId}:
    get:
      tags: [Wallets]
      summary: Get employee wallet balance
      description: Only the wallet owner can fetch their balance.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalance'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /wallet/withdraw/{amount}:
    get:
      tags: [Wallets]
      summary: Withdraw from employee wallet
      description: Only the wallet owner can withdraw funds.
      security:
        - BearerAuth: []
      parameters:
        - name: amount
          in: path
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Withdrawal accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletWithdrawal'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller does not have the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: string
      additionalProperties: false
...
    MainReportData:
      type: object
      properties:
        ordersSum:
          type: number
          format: double
        primeCostSum:
          type: number
          format: double
        earnings:
          type: number
          format: double
        ordersAmount:
          type: integer
          format: int32
        paidOrdersAmount:
          type: integer
          format: int32
        notPaidOrdersAmount:
          type: integer
          format: int32
      required: [ordersSum, primeCostSum, earnings]
    EmployeeReportData:
      type: object
      properties:
        ordersAmount:
          type: integer
          format: int32
        ordersSum:
          type: number
          format: double
        tableAmount:
          type: integer
          format: int32
          description: Count of paid tables
        rating:
          type: number
          format: double
        comments:
          type: array
          items:
            type: string
      required: [ordersAmount, ordersSum]
    EmployeeReportMap:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/EmployeeReportData'
      propertyNames:
        pattern: '^[0-9]+$'
    WalletBalance:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        balance:
          type: number
          format: double
      required: [employeeId, balance]
    WalletWithdrawal:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        amount:
          type: number
          format: double
      required: [employeeId, amount]
...